refactor: separate carousel component from the projects section
